Better modelling of duplications
Have now set the host/species tree as that which defines the duplication
rate, so it's not like parasites speciating independently of the host.
It's a flag so I can (and plan to) go back and do it the old fashioned
way for host/parasite modelling.

Added reporting of number of actual events so I can check how segdup is
doing.

Fixed the timing so parasite/gene nodes can start coevolving with the
host if they're placed *earlier* than the host root.

Segmental / Joint duplication is now achieved by, once a lineage has
been picked at uniform random for duplication, testing each other active
parasite/gene lineage on the same host/species branch against a fixed
probability of hitching a duplication at the same time. Since this is
the same as the model currently in segdup for finding joint duplication
events, this gives us a baseline.